Add repo with result of ingestImages and processCcd
Add a Butler repository that contains the results of running
ingestIMages.py on the data in the 'raw' directory and
processCcd.py in turn on repo with those ingested data.

This is meant to be useful for comparing different processCcd runs
and for providing premade output products to make it easy to test just
one particular stage of processing.